Connect recording flow to Clean Architecture
