+++ DancingSteps/CameraProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  CameraProtocol.swift
+//  DancingSteps
+//
+//  Created by Rodrigo López-Romero Guijarro on 23/08/2017.
+//  Copyright © 2017 RLRG. All rights reserved.
+//
+
+import Foundation
+
+protocol CameraProtocol {
+    func displayError(string: String)
+}
