+++ DancingSteps/DebugViewController.swift
@@ -12,7 +12,7 @@ import MessageUI
 #if DancingStepsLAB
     import FLEX
 #endif
-import Realm // TODO: Remove this from here ? Clean Architecture ?
+import Realm
 
 class DebugViewController: UIViewController, MFMailComposeViewControllerDelegate {
