+++ DancingSteps/UIImageViewExt.swift
@@ -9,7 +9,6 @@
 import Foundation
 import UIKit
 
-// TODO: Refactor this code to Clean Architecture !
 extension UIImageView {
     func downloadedFrom(url: URL, contentMode mode: UIViewContentMode = .scaleAspectFit) {
         contentMode = mode
